removed print statements

+++ src/org/maps/PSO/Particle.java
@@ -2,26 +2,22 @@ package org.maps.PSO;
 
 import org.maps.GA.Chromosome;
 import org.maps.GA.Gene;
-import org.maps.Heft.Heft;
 
-import java.util.HashSet;
 import java.util.Random;
-import java.util.Set;
-import java.util.Vector;
 
 import static org.maps.InputData.Constants.MAX_PROCESSORS;
 import static org.maps.InputData.Constants.MAX_TASKS;
 
 public class Particle extends Chromosome implements Comparable<Particle> {
-    public int[] velocity = new int[MAX_TASKS+1];
+    public float[] velocity = new float[MAX_TASKS + 1];
 
     @Override
     public void generate() {
         Random random = new Random();
-        gene[0] = new Gene(0,0);
+        gene[0] = new Gene(0, 0);
         velocity[0] = -1;
         feasibility = false;
-        while(!feasibility) {
+        while (!feasibility) {
             for (int i = 1; i <= MAX_TASKS; i++) {
                 gene[i] = new Gene(i, random.nextInt(MAX_PROCESSORS) + 1);
                 velocity[i] = random.nextInt(10);
